requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
bbcode>=1.1
Pillow>=10.0
python-decouple>=3.8
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
tldextract>=5.0
geoip2>=4.0                # optional: GeoIP country lookups; needs GeoLite2-Country.mmdb
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
PyJWT>=2.8
